+++ src/gcode/machine/JSONMachine.h
@@ -48,6 +48,7 @@ namespace GCode {
     void setFeedMode(feed_mode_t mode);
     void setSpeed(double speed);
     void setSpinMode(spin_mode_t mode, double max);
+    void setPathMode(path_mode_t mode, double motionBlending, double naiveCAM);
     void changeTool(unsigned tool);
 
     void input(port_t port, input_mode_t mode, double timeout);